+++ src/resourceManager/resourceMessages.js
@@ -1,0 +1,7 @@
+module.exports = {
+    "help": {
+        "usage": "Usage: noddsdale <command>",
+        "errorUnknownCommand": "Unknown command: $(command)",
+        "init": "Usage: noddsdale init/n noddsdale-int purpose is to prepere the current directory for a new noddsdale blog"
+    }
+};